feat: Synchronize profile group properties when updating profile attributes - EXO-60737
Synchronize profile group properties when updating profile attributes
